+++ frontend/src/App.module.css
@@ -1,0 +1,37 @@
+.appContainer {
+    position: relative;
+}
+
+@keyframes fadeInLogo {
+    from {
+        opacity: 0;
+    }
+    to {
+        opacity: 0.4;
+    }
+}
+
+.githubLink {
+    position: absolute;
+    top: 20px;
+    right: 20px;
+    z-index: 1000; 
+    display: block;
+}
+
+.githubLogo {
+    width: 40px; 
+    height: auto; 
+    padding-right: 12px;
+    opacity: 0;
+    transition: opacity 0.2s ease-in-out;
+}
+
+.githubLink.illuminated .githubLogo {
+    opacity: 0.4;
+    animation: fadeInLogo 1.5s ease-in-out;
+}
+
+.githubLink:hover .githubLogo {
+    opacity: 1;
+}
