+++ frontend/src/index.css
@@ -1,4 +1,3 @@
-/* src/index.css */
 body {
     margin: 0;
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
